Squashed commit of the following:

commit 3a30e9111a2d267dd90129887f23d8a6a783a5f0
Author: Anatoly Pilikov <zverre@yandex-team.ru>
Date:   Tue Sep 7 18:03:07 2021 +0300

    Fix overflowing by directly casting to int64

+++ yandex/validators.go
@@ -62,7 +62,7 @@ func ConvertableToInt() schema.SchemaValidateFunc {
 		_, err := strconv.ParseInt(str, 10, 64)
 		if err != nil {
 			es = append(es, fmt.Errorf("expected %s to be an integer in the range (%d, %d), got %q",
-				k, math.MinInt64, math.MaxInt64, str))
+				k, int64(math.MinInt64), int64(math.MaxInt64), str))
 			return
 		}
